Added Many Entities
Added CRUD Registrar
Added College Entity
Added Participant Entity

+++ src/main/java/com/example/eventsmgmt/service/AdminService.java
@@ -81,7 +81,11 @@ public class AdminService {
 	}
 
 	public void deleteById(int id) {
-        userRepository.deleteById(id);
+        User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User Not Found By Id"));
+        Role admin = roleRepository.findByName("ADMIN").orElseThrow(() -> new RuntimeException("Role not found"));
+        if(user.getRole().contains(admin)) {
+            userRepository.deleteById(id);
+        }
 	}
 
 	public void updateAdminById(int id, User user) {
